requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
langgraph
faiss-cpu
rank_bm25
streamlit
sentence-transformers
pandas
requests
watchdog
python-dotenv
PyPDF2
markdown
matplotlib
openmeteo-requests
requests-cache
retry-requests
fastapi
uvicorn[standard]
websockets
pypdf
unstructured
docx2txt
